NEW TRANSITION: cube_rotate
Somewhat mimics a cube's edge rotation towards and then away from the
viewer by adjusting the speed at which the squeeze boundary moves
across the screen. Fastest in the middle and slower near the ends.

Because the wave starts at 45 degrees and ends at 135 degrees
(imagine a square inscribed inside of a unit circle,
with the bottom edge being the face of the cube that is rotating)
the deviation from the normal linear transition in squeeze_expand
is not very noticeable.
There might be a way to exaggerate this mathematically in the future,
but it'll never look quite right because I can't
perspective stretch/squeeze the frames without a custom resize filter,
and because it's not a true projection since that would require the
edges furthest from the viewer become visually smaller,
and this is pretty much not possible with VapourSynth.