Add `exaggeration` param to cube_rotate to adjust speed of transition
Accepts an integer between 0 and 100 inclusive.

The exaggeration represents how much the cosine effect should be
"exaggerated" by acting as a weighted average between
the real projection and a fitted curve.

An `exaggeration` of 0, the default, represents a mathematically correct
projection of the rotation of a face of a cube.
The initial and final velocities (slowest) of the transition boundary,
when compared to the speed of the linear transition in squeeze_expand,
are ~0.785x as fast. The speed at the midpoint (fastest) of the
transition is ~1.11x as fast

An `exaggeration` of 100 represents a fitted cosine wave.
The initial and final velocities (slowest) of the transition boundary,
when compared to the linear transition in squeeze_expand,
are 0.00x as fast (derivative of the position wrt time is 0).
At the midpoint (fastest) of the transition, the speed is
~1.571x as fast.
Full `100` exaggeration looks a bit more "cubic" than the normal
rotation because it looks more like a wide-angle lens effect
(closer to real perspective and not a flat projection onto 2D space).